backend_ai_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.3
google-generativeai==0.3.2
pypdf
python-docx==1.1.0
python-multipart==0.0.6
numpy==1.24.3
python-dotenv==1.0.0
pgvector==0.2.5
websockets==12.0